requirements.txt
streamlit
pymongo
PyPDF2
sentence-transformers
pandas
numpy
scikit-learn
python-dotenv
